+++ src/routes/+page.svelte
@@ -1,8 +1,19 @@
 <script lang="ts">
+  import { page } from '$app/stores';
   import type { ActionData, PageData } from './$types';
 
   export let data: PageData;
   export let form: ActionData;
+
+  let log: string[] = [];
+
+  async function preflight(method: 'POST' | 'GET') {
+    const response = await fetch(new URL('/preflight-required', $page.url), {
+      method
+    });
+    const json = await response.json();
+    log = [...log, json.message];
+  }
 </script>
 
 <h1>Try the rate limiter</h1>
@@ -24,6 +35,15 @@
   <button>Submit</button>
 </form>
 
+<hr />
+
+<button on:click={() => preflight('POST')}>POST to preflight route</button>
+<button on:click={() => preflight('GET')}>GET preflight route</button>
+
+<pre>
+{#each log as msg}{msg}<br />{/each}
+</pre>
+
 <style>
   form {
     font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI',
@@ -35,6 +55,10 @@
     padding: 0.5rem;
     color: white;
     background-color: forestgreen;
+  }
+
+  hr {
+    margin: 2rem 0;
   }
 
   .limited {